requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
-f https://download.openmmlab.com/mmcv/dist/cu117/torch1.13.0/index.html

addict==2.4.0
attrs==22.1.0
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
Cython<3
exceptiongroup==1.0.4
fonttools==4.38.0
ftfy==6.1.1
idna==3.4
iniconfig==1.1.1
kiwisolver==1.4.4
lanms-neo==1.0.2
lazy_loader==0.4
lmdb==1.4.1
matplotlib==3.6.2
mmcls==0.25.0
mmcv-full==1.7.0
mmdet==2.26.0
mmocr==0.6.3
mmsegmentation==0.30.0
networkx==3.1
numpy==1.23.5
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pluggy==1.0.0
prettytable==3.11.0
pyclipper==1.3.0.post6
pycocotools==2.0.7
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==3.9.7
regex==2022.10.31
requests==2.28.1
scikit-image==0.21.0
scipy==1.9.3
Shapely==2.0.0
six==1.16.0
terminaltables==3.1.10
tifffile==2023.7.10
timm==1.0.9
tomli==2.0.1
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.13
yapf==0.32.0
